easyback/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
markdown
django-filter
gerencianet
requests
sqlparse
djoser